+++ src/pages/Cart/Cart.js
@@ -1,4 +1,5 @@
-import React from 'react';
+import { React, useState, useEffect } from 'react';
+import CartBox from './component/CartBox';
 import './Cart.scss';
 
 const Cart = () => {
@@ -36,27 +37,3 @@ const Cart = () => {
 };
 
 export default Cart;
-
-const CartBox = () => {
-  return (
-    <section className="cart-box">
-      <div className="cart-img">
-        <img
-          src="/images/ProductDetail/plant-21.jpg"
-          alt="product-img"
-          className="img"
-        />
-      </div>
-      <div className="cart-content">
-        <div className="content">
-          <h1 className="title">몬스테라</h1>
-          <p>수량 : 2개</p>
-          <button className="btn">삭제</button>
-        </div>
-        <div className="total">
-          <p>총 가격:20,000원</p>
-        </div>
-      </div>
-    </section>
-  );
-};
