+++ src/pages/Cart/Cart.scss
@@ -21,47 +21,6 @@ button:hover {
   .cart-main {
     width: 60%;
     margin-right: 40px;
-
-    .cart-box {
-      display: flex;
-      padding: 25px 0 20px 0;
-      border-bottom: 1px solid #dfdfdf;
-
-      .cart-img {
-        width: 35%;
-        margin-right: 10px;
-      }
-
-      .btn {
-        padding: 0;
-        border: none;
-        background-color: white;
-        font-size: 20px;
-      }
-
-      .cart-content {
-        display: flex;
-        justify-content: space-between;
-        width: 100%;
-        padding: 10px;
-      }
-
-      .content {
-        display: flex;
-        flex-direction: column;
-        align-items: flex-start;
-        justify-content: space-between;
-      }
-
-      .img {
-        width: 100%;
-      }
-
-      .total {
-        display: flex;
-        flex-direction: column-reverse;
-      }
-    }
   }
 
   .cart-aside {
@@ -69,7 +28,7 @@ button:hover {
     padding-top: 25px;
 
     .price {
-      padding: 10px 0 10px 0;
+      padding: 10px 0;
       border-bottom: 1px solid #dfdfdf;
     }
