+++ src/pages/Cart/component/CartBox.scss
@@ -1,0 +1,40 @@
+.cart-box {
+  display: flex;
+  padding: 25px 0 20px 0;
+  border-bottom: 1px solid #dfdfdf;
+
+  .cart-img {
+    width: 35%;
+    margin-right: 10px;
+  }
+
+  .btn {
+    padding: 0;
+    border: none;
+    background-color: white;
+    font-size: 20px;
+  }
+
+  .cart-content {
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+    padding: 10px;
+  }
+
+  .content {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    justify-content: space-between;
+  }
+
+  .img {
+    width: 100%;
+  }
+
+  .total {
+    display: flex;
+    flex-direction: column-reverse;
+  }
+}
